Use 8 bytes for RNG(Handle) seed on both platforms.
  Change seed type from unsigned long (4 bytes on windows, 8 bytes on
  linux) to uint64_t. Bump cnkt model version to 17 (v1). Change a few
  int64_t to uint64_t along the way. Close #1124.

+++ Source/Math/NoGPU.cpp
@@ -2241,7 +2241,7 @@ void GPUDataTransferer::WaitForCopyCPUToGPUAsync(){}
 
 #pragma region GPURNGHandle functions
 
-GPURNGHandle::GPURNGHandle(int deviceId, unsigned long seed, unsigned long long offset)
+GPURNGHandle::GPURNGHandle(int deviceId, uint64_t seed, uint64_t offset)
     : RNGHandle(deviceId)
 {
 }
